MBS-12392 Fix teamcity metrics
