+++ subprojects/assemble/build-metrics/build.gradle.kts
@@ -16,6 +16,7 @@ dependencies {
     implementation(projects.subprojects.gradle.gradleExtensions)
     implementation(projects.subprojects.gradle.statsdConfig)
     implementation(projects.subprojects.gradle.teamcity)
+    implementation(projects.subprojects.gradle.worker)
     implementation(libs.kotlinGradle)
 
     testImplementation(libs.mockitoKotlin)
